+++ docker-compose.yml
@@ -25,7 +25,7 @@ x-airflow-common:
     - ./dags:/opt/airflow/dags
     - ./plugins:/opt/airflow/plugins
     - ./sm2a-local-config/local_airflow.cfg:/opt/airflow/airflow.cfg
-    - ./airflow_services/webserver_config_keycloak.py:/opt/airflow/webserver_config.py
+    - ./airflow_services/webserver_config.py:/opt/airflow/webserver_config.py
     - ./infrastructure/configuration:/opt/airflow/configuration
     - ./scripts:/opt/airflow/scripts
   user: "50000:0"
